shell sort error's fixed

+++ 100-shell_sort.c
@@ -28,7 +28,7 @@
 
 void shell_sort(int *array, size_t size)
 {
-	unsigned int i = 0, j = 0, gap = 0;
+	size_t i = 0, j = 0, gap = 0;
 	int aux = 0;
 
 	if (array == NULL || size < 2)
